Added issue template for CMIP6 CMORization and ESGF-publication request
Special issue template for NorESM data providers that want to CMORize and ESGF-publish a CMIP6 experiment.

The data provider creates the issue and the person/team (currently @YanchunHe) who deals with the CMORization is automatically assigned. 

Once the CMORization and QC is completed, a commented should be posted with a mention of the person/team (currently @monsieuralok) who deals with the ESGF-publication. 

The issues will be closed after successful ESGF publication and can be re-opened in case of retraction & data correction.